+++ app/build.gradle
@@ -54,17 +54,17 @@ android {
 }
 
 dependencies {
-    implementation 'androidx.core:core-ktx:1.5.0'
+    implementation 'androidx.core:core-ktx:1.6.0'
     implementation 'androidx.appcompat:appcompat:1.3.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
     implementation 'androidx.viewpager2:viewpager2:1.0.0'
     implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.3.1'
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.3.1'
-    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.3.1'
+    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.4.0-alpha02'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    implementation 'androidx.datastore:datastore-preferences:1.0.0-beta02'
-    implementation "androidx.fragment:fragment-ktx:1.3.5"
-    implementation 'com.google.android.material:material:1.3.0'
+    implementation 'androidx.datastore:datastore-preferences:1.0.0-rc01'
+    implementation 'androidx.fragment:fragment-ktx:1.3.5'
+    implementation 'com.google.android.material:material:1.4.0'
     implementation 'com.github.chrisbanes:PhotoView:2.3.0'
     implementation 'com.google.zxing:core:3.3.3'
     implementation 'org.bouncycastle:bcprov-jdk15to18:1.68'
