+++ app/src/main/java/com/michaeltroger/gruenerpass/ViewEvent.kt
@@ -1,0 +1,8 @@
+package com.michaeltroger.gruenerpass
+
+sealed class ViewEvent {
+    object ShowPasswordDialog : ViewEvent()
+    object ShowReplaceDialog : ViewEvent()
+    object ShowDeleteDialog : ViewEvent()
+    object CloseAllDialogs : ViewEvent()
+}
