+++ app/src/main/java/com/michaeltroger/gruenerpass/ViewState.kt
@@ -1,0 +1,7 @@
+package com.michaeltroger.gruenerpass
+
+sealed class ViewState {
+    object Empty : ViewState()
+    object Certificate : ViewState()
+    object Error : ViewState()
+}
